⚡

+++ src/interactive/repl.ts
@@ -60,17 +60,15 @@ async function confirmKill() {
 }
 async function stopREPL(onDeactivate=false) {
     const config = vscode.workspace.getConfiguration('julia')
-    if (Boolean(config.get('persistentSession.enabled'))) {
-        if (!onDeactivate) {
-            try {
-                const sessionName = parseSessionArgs(config.get('persistentSession.tmuxSessionName'))
-                const killSession = await confirmKill()
-                if (killSession) {
-                    await exec(`tmux kill-session -t ${sessionName}`)
-                }
-            } catch (err) {
-                vscode.window.showErrorMessage('Failed to close tmux session: ' + err.stderr)
+    if (Boolean(config.get('persistentSession.enabled')) && !onDeactivate) {
+        try {
+            const sessionName = parseSessionArgs(config.get('persistentSession.tmuxSessionName'))
+            const killSession = await confirmKill()
+            if (killSession) {
+                await exec(`tmux kill-session -t ${sessionName}`)
             }
+        } catch (err) {
+            vscode.window.showErrorMessage('Failed to close tmux session: ' + err.stderr)
         }
     }
     if (isConnected()) {
